requirements.txt
absl-py==0.15.0
asn1crypto==0.24.0
astor==0.8.1
atomicwrites==1.1.5
attrs==18.2.0
backports.functools-lru-cache==1.5
backports.shutil-get-terminal-size==1.0.0
backports.weakref==1.0rc1
beautifulsoup4==4.7.1
bleach==1.5.0
chardet==3.0.4
click==7.1.2
configparser==3.5.0b2
cryptography==2.6.1
cvxopt==1.1.9
cycler==0.10.0
decorator==4.3.0
entrypoints==0.3
enum34==1.1.6
et-xmlfile==1.0.1
funcsigs==1.0.2
future==0.16.0
futures==3.3.0
gast==0.5.3
gmpy==1.17
google-pasta==0.2.0
grpcio==1.41.1
h5py==2.10.0
html5lib==0.9999999
iniparse==0.4
ipaddress==1.0.17
ipython==5.8.0
ipython-genutils==0.2.0
jdcal==1.0
joblib==0.13.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==17.1.1
keyrings.alt==3.1.1
kiwisolver==1.0.1
lxml==4.3.2
lz4==1.1.0
Markdown==3.1.1
matplotlib==2.2.3
mock==2.0.0
more-itertools==4.2.0
mpmath==1.1.0
mysqlclient==1.3.10
nose==1.3.7
numexpr==2.6.9
numpy==1.16.2
olefile==0.46
openpyxl==2.4.9
pandas==0.23.3+dfsg
pathlib2==2.3.3
patsy==0.5.0+dev
pbr==4.2.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
pluggy==0.8.0
prompt-toolkit==1.0.15
protobuf==3.17.3
psutil==5.5.1
py==1.7.0
pycrypto==2.6.1
pycurl==7.43.0.2
pyglet==1.3.0
Pygments==2.3.1
PyGObject==3.30.4
pyliblzma==0.5.3
pyparsing==2.2.0
pysqlite==1.0.1
pytest==3.10.1
python-dateutil==2.7.3
pytz==2019.1
pyxdg==0.25
PyYAML==5.4.1
rpm==4.14.2.1
scandir==1.9.0
scipy==1.1.0
SecretStorage==2.3.1
simplegeneric==0.8.1
simplejson==3.16.0
six==1.12.0
soupsieve==1.8
statsmodels==0.8.0
subprocess32==3.5.3
sympy==1.3
tables==3.4.4
tensorboard==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
Theano==1.0.5
traitlets==4.3.2
urlgrabber==3.10.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.13.3
xlrd==1.1.0
xlwt==1.3.0
yum-metadata-parser==1.1.4
